Django/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
asgiref=3.4.1=pypi_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.10.0=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39h06a4308_0
cffi=1.14.6=py39h400218f_0
charset-normalizer=2.0.9=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pypi_0
coreapi=2.3.3=pypi_0
coreschema=0.0.4=pypi_0
cryptography=35.0.0=py39hd23ed53_0
cuda80=1.0=h205658b_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.25=pypi_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.0=pyhd8ed1ab_0
detectron2=0.6+cu111=pypi_0
django=3.2.6=pypi_0
djangorestframework=3.12.4=pypi_0
drf-yasg=1.20.0=pypi_0
elasticsearch=7.16.0=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3=hf484d3e_0
flatbuffers=1.12=pypi_0
fonttools=4.28.3=pypi_0
freetype=2.11.0=h70c0345_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20211023=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-api-core=2.3.0=pypi_0
google-api-python-client=2.33.0=pypi_0
google-auth=2.3.3=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-cloud-vision=2.6.2=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.54.0=pypi_0
grpcio=1.34.1=pypi_0
grpcio-status=1.42.0=pypi_0
gunicorn=20.1.0=pypi_0
h5py=3.1.0=pypi_0
httplib2=0.20.2=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=pypi_0
inflection=0.5.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
ipykernel=6.6.0=py39hef51801_0
ipython=7.30.1=py39hf3d152e_0
itypes=1.2.0=pypi_0
jedi=0.18.1=py39hf3d152e_0
jinja2=3.0.3=pypi_0
jpeg=9d=h7f8727e_0
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py39hf3d152e_1
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
numpy=1.21.4=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.1.1=pypi_0
opencv-python=4.5.4.60=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.9.0=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
portalocker=2.3.2=pypi_0
prompt-toolkit=3.0.24=pyha770c72_0
proto-plus=1.19.8=pypi_0
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.3=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.6=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py39hb69f2a1_2
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
ruamel-yaml=0.17.17=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
setuptools=58.0.4=py39h06a4308_0
setuptools-scm=6.3.2=pypi_0
six=1.15.0=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.36.0=hc218d9a_0
sqlparse=0.4.2=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=1.2.2=pypi_0
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=py39_cu113
tornado=6.1=py39h3811e60_1
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2021e=hda174b7_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
